+++ lib/sigaa_parser/transcript_parser.rb
@@ -29,6 +29,7 @@ module SigaaParser
 
       parse_student
       parse_results
+      parse_progress
     end
 
   protected
@@ -52,6 +53,16 @@ module SigaaParser
 
         @course_results << result if result
       end
+    end
+
+    def parse_progress
+      total = @rows.find { |line| line.include?('Exigido') }.split(' ')
+      completed = @rows.find { |line| line.include?('Integralizad') }.split(' ')
+
+      # "Básicas Profissionais" + "Complementares obrigatórias"
+      @course_results.progress.compulsory.set(completed[1].to_i + completed[10].to_i, total[1].to_i + total[10].to_i)
+      @course_results.progress.optional.set(completed[4].to_i, total[4].to_i)
+      @course_results.progress.flexible.set(completed[7].to_i, total[7].to_i)
     end
 
     def parse_row_result(row)