+++ spec/transcript_parser_spec.rb
@@ -1,17 +1,16 @@
 describe SigaaParser::TranscriptParser do
   let(:file) { File.open(File.join(TRANSCRIPT_DIR, '11111309.pdf')) }
-  subject { SigaaParser::TranscriptParser.new(file) }
+  subject { SigaaParser::TranscriptParser.new(file).course_results }
 
   describe '#new' do
     context 'when file is a PDF and a transcript of records' do
       it 'finds student data' do
-        expect(subject.course_results.student.name).to be_eql('FERNANDO SANTOS DE MATTOS BRITO')
-        expect(subject.course_results.student.program).to be_eql('CIÊNCIAS DA COMPUTAÇÃO (BACHARELADO) - CI/João Pessoa')
+        expect(subject.student.name).to be_eql('FERNANDO SANTOS DE MATTOS BRITO')
+        expect(subject.student.program).to be_eql('CIÊNCIAS DA COMPUTAÇÃO (BACHARELADO) - CI/João Pessoa')
       end
 
-
       it 'finds all courses' do
-        expect(subject.course_results.results.size).to be_eql(52)
+        expect(subject.results.size).to be_eql(52)
       end
     end
 
@@ -32,5 +31,16 @@ describe SigaaParser::TranscriptParser do
     end
   end
 
+  describe "#results" do
+    it 'parses the data correctly' do
+      expect(subject.progress.compulsory.completed).to be(97)
+      expect(subject.progress.compulsory.total).to be(190)
 
+      expect(subject.progress.optional.completed).to be(14)
+      expect(subject.progress.optional.total).to be(16)
+
+      expect(subject.progress.flexible.completed).to be(4)
+      expect(subject.progress.flexible.total).to be(12)
+    end
+  end
 end